Bug fix: Halt run if Restart.frequency and Restart.duration differ
We have put in an error check in History/history_mod.F90 that will
halt any simulation if the Restart.frequency and Restart.duration
values differ.

NOTE: The algorithm checks if FileWriteYmd == FileCloseYmd and if
FileWriteHms == FileCloseHms.  So if you have a 1 day simulation and
Restart.frequency = 00000001 000000 and Restart.duration = 'End',
the error check will allow the run to proceed.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>